general_question_answering/requirements.txt
google-generativeai>=0.3.0
yfinance>=0.2.36
python-dotenv>=1.0.0
typing-extensions>=4.8.0
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
pymongo>=4.6.0
tqdm>=4.66.0
